working with visual studio version

+++ Source/SolverFraser.cpp
@@ -7,10 +7,10 @@ namespace SPH {
 // 2. Calc a(t+dt) using x(t+dt)
 // 3. v(t+dt) = v(t) dt + 1/2 (a(t) +a(t+dt)) * dt
 ///////////////////////////////////////
-inline void Domain::SolveDiffUpdateFraser (double tf, double dt, double dtOut, char const * TheFileKey, int maxidx) {
+inline void Domain::SolveDiffUpdateFraser (double tf, double dt, double dtOut, char const * TheFileKey, size_t maxidx) {
 	std::cout << "\n--------------Solving---------------------------------------------------------------" << std::endl;
 
-	int idx_out = 1;
+	size_t idx_out = 1;
 	double tout = Time;
 
 	//Initializing adaptive time step variables
@@ -208,7 +208,7 @@ inline void Domain::SolveDiffUpdateFraser (double tf, double dt, double dtOut, c
       DensReduction();
     //#endif
     #pragma omp parallel for schedule (static) num_threads(Nproc)
-    for (int i=0; i<Particles.Size(); i++){
+    for (size_t i=0; i<Particles.Size(); i++){
       //Particles[i]->UpdateDensity_Leapfrog(deltat);
       Particles[i]->Density += deltat*Particles[i]->dDensity;
     }    
@@ -223,7 +223,7 @@ inline void Domain::SolveDiffUpdateFraser (double tf, double dt, double dtOut, c
     RateTensorsReduction();
     //#endif
     #pragma omp parallel for schedule (static) num_threads(Nproc)
-    for (int i=0; i<Particles.Size(); i++){
+    for (size_t i=0; i<Particles.Size(); i++){
       //Particles[i]->Mat2Leapfrog(deltat); //Uses density  
       Particles[i]->CalcStressStrain(deltat); //Uses density  
     } 
@@ -259,7 +259,7 @@ inline void Domain::SolveDiffUpdateFraser (double tf, double dt, double dtOut, c
     //14. Add contact
     // if (contact ){
     // #pragma omp parallel for schedule (static) num_threads(Nproc)
-    // for (int i=0; i<Particles.Size(); i++)
+    // for (size_t i=0; i<Particles.Size(); i++)
       // Particles[i]->a += Particles[i] -> contforce / Particles[i] -> Mass; 
     // }
     //// UPDATE VEL AND POS
@@ -270,7 +270,7 @@ inline void Domain::SolveDiffUpdateFraser (double tf, double dt, double dtOut, c
     //CorrectVelAcc();
     MoveGhost(); 
    #pragma omp parallel for schedule (static) private(du) num_threads(Nproc)
-    for (int i=0; i<Particles.Size(); i++){
+    for (size_t i=0; i<Particles.Size(); i++){
       Particles[i]->x_prev = Particles[i]->x;
       du = (Particles[i]->v + Particles[i]->VXSPH)*deltat + 0.5 * Particles[i]->a *deltat*deltat;
       Particles[i]->Displacement += du;
@@ -285,7 +285,7 @@ inline void Domain::SolveDiffUpdateFraser (double tf, double dt, double dtOut, c
     clock_beg = clock();
 
     #pragma omp parallel for schedule (static) num_threads(Nproc)
-    for (int i=0; i<Particles.Size(); i++){
+    for (size_t i=0; i<Particles.Size(); i++){
       Particles[i]->v += Particles[i]->a * deltat;
     }
     //CorrectVelAcc();
@@ -295,7 +295,7 @@ inline void Domain::SolveDiffUpdateFraser (double tf, double dt, double dtOut, c
     mov_time_spent += (double)(clock() - clock_beg) / CLOCKS_PER_SEC;  
 
     #pragma omp parallel for schedule (static) num_threads(Nproc)
-    for (int i=0; i<Particles.Size(); i++)
+    for (size_t i=0; i<Particles.Size(); i++)
       prev_acc[i] = Particles[i]->a;
 
 		CalcPlasticWorkHeat(deltat);   //Before Thermal increment because it is used
@@ -338,7 +338,7 @@ inline void Domain::SolveDiffUpdateFraser (double tf, double dt, double dtOut, c
 		}
     
     if (Particles[0]->FirstStep)
-    for (int i=0; i<Particles.Size(); i++){
+    for (size_t i=0; i<Particles.Size(); i++){
       Particles[i]->FirstStep = false;
     }
 		if (isfirst) isfirst = false;
